% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_check.R
\name{Check_base_info}
\alias{Check_base_info}
\title{Check the format of sheet 'base'}
\usage{
Check_base_info(input, dataset_format = 1, as_EN_colnames = TRUE)
}
\arguments{
\item{input}{input}

\item{dataset_format}{dataset_format (default as 1)}

\item{as_EN_colnames}{Convert colnames to EN format (default as TRUE)}
}
\value{
List
}
\description{
Check the format of sheet 'base'
}
