% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_check.R
\name{Check_format}
\alias{Check_format}
\title{Main function to check the spread format using Check_sheet_name, Check_sample_id,
  and Check_base_info.}
\usage{
Check_format(fn, dataset_format = 1, as_EN_colnames = TRUE)
}
\arguments{
\item{fn}{Filename}

\item{dataset_format}{The format of dataset to be checked.}

\item{as_EN_colnames}{Convert colnames to EN format (default as \code{TRUE})}
}
\value{
A list
}
\description{
Main function to check the spread format using Check_sheet_name, Check_sample_id,
  and Check_base_info.
}
