% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality_control.R
\name{Check_geo_location}
\alias{Check_geo_location}
\title{Check the geo-location through heatmap}
\usage{
Check_geo_location(input)
}
\arguments{
\item{input}{input}
}
\value{
List
}
\description{
Check the geo-location through heatmap
}
