% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMATO.R
\name{Check_param_quality}
\alias{Check_param_quality}
\title{Check the quality of parameters}
\usage{
Check_param_quality(input)
}
\arguments{
\item{input}{input}
}
\value{
List
}
\description{
Check the quality of parameters
}
