% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMATO.R
\name{Check_sheet_name}
\alias{Check_sheet_name}
\title{Check the format of sheet names of the spread}
\usage{
Check_sheet_name(fn, dataset_format = 1)
}
\arguments{
\item{fn}{fn}

\item{dataset_format}{dataset_format (default as 1)}
}
\value{
List
}
\description{
Check the format of sheet names of the spread
}
