% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quality_control.R
\name{Check_spectra_quality}
\alias{Check_spectra_quality}
\title{Check the quality of spectra}
\usage{
Check_spectra_quality(input)
}
\arguments{
\item{input}{input}
}
\value{
List
}
\description{
Check the quality of spectra
}
