% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_check.R
\name{Check_spread_file}
\alias{Check_spread_file}
\title{Check the readable of spread files}
\usage{
Check_spread_file(fn)
}
\arguments{
\item{fn}{fn}
}
\value{
Logical
}
\description{
Check the readable of spread files
}
