% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_control.R
\name{cli_control}
\alias{cli_control}
\title{Control the position and information should be printed}
\usage{
cli_control(pos = "info", text, col_status = NULL)
}
\arguments{
\item{pos}{The position of cli function should be}

\item{text}{The character to print}

\item{col_status}{The status color for pos = end}
}
\description{
Control the position and information should be printed
}
\examples{
cli_control("head", "Message 1")
cli_control("okay", "Okay message")
cli_control("warning", "Warning message")
cli_control("error", "Error message")
cli_control("info", "Plain information")
cli_control("end", "End of message")

}
