% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMATO.R
\docType{data}
\encoding{UTF-8}
\name{dataset_format_1}
\alias{dataset_format_1}
\title{Format_1 used for generating reference spreads}
\usage{
dataset_format_1
}
\description{
Format_1 used for generating reference spreads
}
