% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_collection.R
\docType{data}
\encoding{UTF-8}
\name{dataset_format_1_folders}
\alias{dataset_format_1_folders}
\title{Format_1 used for generating reference folders}
\usage{
dataset_format_1_folders
}
\description{
Format_1 used for generating reference folders
}
