% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_template.R
\name{Generate_ref_folders}
\alias{Generate_ref_folders}
\title{Generate_ref_folders}
\usage{
Generate_ref_folders(dataset_format = 1)
}
\arguments{
\item{dataset_format}{dataset_format}
}
\value{
The created templated folders shoud be at the working directory.
}
\description{
Generate_ref_folders
}
