% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMATO.R
\name{Generate_ref_spread}
\alias{Generate_ref_spread}
\title{Generate spreads of the built-in format (recommand)}
\usage{
Generate_ref_spread(dataset_format = 1, file_format = "xlsx")
}
\arguments{
\item{dataset_format}{dataset_format (default as 1)}

\item{file_format}{file_format (default as 'xlsx')}
}
\value{
List
}
\description{
Generate spreads of the built-in format (recommand)
}
