% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Render_report.R
\name{render_report}
\alias{render_report}
\title{render_report}
\usage{
render_report(input, RMD = NULL, output_format = NULL, ...)
}
\arguments{
\item{input}{Input list}

\item{RMD}{RMD file path}

\item{output_format}{Param in rmarkdown::render, try \code{"html_document"},
\code{"rmdformats::readthedown"}, \code{"pagedown::html_paged"}, or 
\code{"slidy_presentation"}.}

\item{...}{...}
}
\description{
render_report
}
