% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMATO.R
\name{Transpose_spec_df}
\alias{Transpose_spec_df}
\title{Transpose the spectra data as a data.frame with wavelength being colnames}
\usage{
Transpose_spec_df(input, dataset_format = 1)
}
\arguments{
\item{input}{Result of Check_base_info}

\item{dataset_format}{Dataset_format (default as 1)}
}
\value{
A list or NULL (error output)
}
\description{
Transpose the spectra data as a data.frame with wavelength being colnames
}
