% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAMATO.R
\encoding{UTF-8}
\name{Check_sample_id}
\alias{Check_sample_id}
\title{Check the format of sample ids of the spread}
\usage{
Check_sample_id(input, dataset_format = 1)
}
\arguments{
\item{input}{input}

\item{dataset_format}{dataset_format (default as 1)}
}
\value{
List
}
\description{
Check the format of sample ids of the spread
}
