% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_template.R
\name{Generate_ref_spread}
\alias{Generate_ref_spread}
\title{Generate spreads of the built-in format (recommand)}
\usage{
Generate_ref_spread(
  dataset_format = 1,
  file_format = "xlsx",
  overwrite = FALSE,
  use_CN = TRUE,
  QF_sheet = FALSE,
  Meta_sheet = TRUE,
  add_demo = TRUE,
  add_unit = TRUE
)
}
\arguments{
\item{dataset_format}{dataset_format (default as 1)}

\item{file_format}{file_format (default as 'xlsx')}

\item{overwrite}{Default as \code{FALSE}, just in case you forget to copy demo spread to other folders.}

\item{use_CN}{Use CN colnames (default as \code{TRUE})}

\item{QF_sheet}{Qualtiy Flag sheets (default as \code{FALSE})}

\item{Meta_sheet}{Metadata info sheets (default as \code{TRUE})}

\item{add_demo}{Add demo samples (TRUE)}

\item{add_unit}{Add unit (TRUE)}
}
\value{
List
}
\description{
Generate spreads of the built-in format (recommand)
}
